+++ src/components/controls/NodeSelector.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import Dropdown from './Dropdown2.vue'
+import Dropdown from './Dropdown.vue'
 export default {
   components: {
     Dropdown
